Refactor, cleanup, organize project files, separate listeners. #7
