+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: TorchLight
-main: com.stdnullptr.Main
+main: com.stdnullptr.PluginMain
 version: 1.1.1
 api-version: 1.19
 description: The plugin makes the torch emit light when held in offhand
